requirements.txt
Flask
Flask-Bootstrap
numpy
opencv-python
requests
pyaudio
SpeechRecognition
pydub
torch
tqdm
pillow
whisper
editdistance
rlvoice-1
git+https://github.com/openai/whisper.git
--extra-index-url https://download.pytorch.org/whl/cu113
torch
openai
tqdm
pyserial>=2.7